requirements.txt
albumentations==1.3.1
dotdict==0.1
imutils==0.5.4
ipykernel==6.15.1
ipywidgets==8.0.1
keras==2.11.0
matplotlib==3.5.2
numpy==1.24.4
opencv-python==4.6.0.66
openpyxl==3.0.10
pandas==1.4.3
Pillow==9.2.0
pydicom==2.3.0
PyYAML==6.0
ray==2.6.1
scikit-image==0.21.0
scikit-learn==1.1.2
scipy==1.9.0
seaborn==0.12.1
setuptools==59.5.0
SimpleITK==2.4.0
tensorboard==2.11.2
tensorboardX==2.6.2
tensorflow==2.11.0
timm==0.6.12
torch==1.10.0+cu113
-f https://download.pytorch.org/whl/torch_stable.html
torchsummary==1.5.1
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.11.0+cu113
-f https://download.pytorch.org/whl/torch_stable.html
tqdm==4.64.0